Implementation of A* Search Algorithm for a Rolling Dice on a Maze
We implement A* search in Python to solve rolling-die mazes. In a rolling-die maze, the objective is to"roll" a die along its edges through a grid until a goal location is reached. The initial state of the search isgiven by the die location and the orientation of the die faces. Rolling die mazes contain obstacles, along withrestrictions on which numbers may face 'up.'